Unify types/mod.rs and value.rs
